+++ server/public.go
@@ -969,7 +969,7 @@ func (s *PublicServer) explorerSearch(w http.ResponseWriter, r *http.Request) (t
 		findAssets = s.api.FindAssets(q, 0, 2)
 		if len(findAssets.AssetDetails) > 0 {
 			if len(findAssets.AssetDetails) == 1 {
-				http.Redirect(w, r, joinURL("/asset/", strconv.FormatUint(uint64(findAssets.AssetDetails[0].AssetGuid), 10)), 302)
+				http.Redirect(w, r, joinURL("/asset/", strconv.FormatUint(findAssets.AssetDetails[0].AssetGuid) 10)), 302)
 				return noTpl, nil, nil
 			} else {
 				http.Redirect(w, r, joinURL("/assets/", q), 302)
@@ -978,7 +978,7 @@ func (s *PublicServer) explorerSearch(w http.ResponseWriter, r *http.Request) (t
 		}
 		asset, err = s.api.GetAsset(q, 0, 1, api.AccountDetailsBasic, &api.AddressFilter{AssetsMask: bchain.AssetMask})
 		if err == nil {
-			http.Redirect(w, r, joinURL("/asset/", strconv.FormatUint(uint64(asset.AssetDetails.AssetGuid), 10)), 302)
+			http.Redirect(w, r, joinURL("/asset/", strconv.FormatUint(asset.AssetDetails.AssetGuid, 10)), 302)
 			return noTpl, nil, nil
 		}
 	}
